+++ medusa/name_parser/parser.py
@@ -5,7 +5,6 @@ from __future__ import unicode_literals
 
 import logging
 import time
-from builtins import object
 from collections import OrderedDict
 
 import guessit
@@ -62,7 +61,7 @@ class NameParser(object):
         airdate = result.air_date.toordinal()
         main_db_con = db.DBConnection()
         sql_result = main_db_con.select(
-            b'SELECT season, episode FROM tv_episodes WHERE indexer = ? AND showid = ? AND airdate = ?',
+            'SELECT season, episode FROM tv_episodes WHERE indexer = ? AND showid = ? AND airdate = ?',
             [result.series.indexer, result.series.series_id, airdate])
 
         return sql_result
@@ -88,14 +87,14 @@ class NameParser(object):
         episode_numbers = []
 
         if episode_by_air_date:
-            season_number = int(episode_by_air_date[0][0])
-            episode_numbers = [int(episode_by_air_date[0][1])]
+            season_number = int(episode_by_air_date[0]['season'])
+            episode_numbers = [int(episode_by_air_date[0]['episode'])]
 
             # Use the next query item if we have multiple results
             # and the current one is a special episode (season 0)
             if season_number == 0 and len(episode_by_air_date) > 1:
-                season_number = int(episode_by_air_date[1][0])
-                episode_numbers = [int(episode_by_air_date[1][1])]
+                season_number = int(episode_by_air_date[1]['season'])
+                episode_numbers = [int(episode_by_air_date[1]['episode'])]
 
             log.debug(
                 'Database info for series {name}: Season: {season} Episode(s): {episodes}', {
@@ -503,6 +502,9 @@ class ParseResult(object):
                                              quality=common.Quality.qualityStrings[self.quality],
                                              total_time=self.total_time))
         return helpers.canonical_name(obj, fmt='{key}: {value}', separator=', ')
+
+    # Python 2 compatibility
+    __unicode__ = __str__
 
     def get_quality(self, guess, extend=False):
         """Return video quality from guess or name.