+++ medusa/server/api/v2/alias_source.py
@@ -16,7 +16,7 @@ def find_alias_sources(predicate=None):
     mapping = {'local': 'custom_exceptions'}
     for identifier in ('local', 'xem', 'anidb'):
         if not predicate or predicate(identifier):
-            last_refresh = get_last_refresh(mapping.get(identifier, identifier))[0][b'last_refreshed']
+            last_refresh = get_last_refresh(mapping.get(identifier, identifier))[0]['last_refreshed']
             data.append({'id': identifier, 'lastRefresh': last_refresh})
 
     return data
@@ -34,7 +34,7 @@ class AliasSourceHandler(BaseRequestHandler):
     #: allowed HTTP methods
     allowed_methods = ('GET', )
 
-    def get(self, identifier, path_param=None):
+    def http_get(self, identifier, path_param=None):
         """Query alias source information.
 
         :param identifier: source name
@@ -71,7 +71,7 @@ class AliasSourceOperationHandler(BaseRequestHandler):
     #: allowed HTTP methods
     allowed_methods = ('POST', )
 
-    def post(self, identifier):
+    def http_post(self, identifier):
         """Refresh all scene exception types."""
         types = {
             'local': 'custom_exceptions',