+++ medusa/show/recommendations/trakt.py
@@ -165,7 +165,7 @@ class TraktPopular(object):
                             continue
                     else:
                         trending_shows.append(self._create_recommended_show(
-                            show, storage_key=b'trakt_{0}'.format(show['show']['ids']['trakt'])
+                            show, storage_key='trakt_{0}'.format(show['show']['ids']['trakt'])
                         ))
 
                 except MultipleShowObjectsException: