+++ tests/providers/test_generic_provider.py
@@ -2,7 +2,7 @@
 """Provider test code for Generic Provider."""
 from __future__ import unicode_literals
 
-from datetime import date, datetime
+from datetime import date, datetime, timedelta
 
 from dateutil import tz
 
@@ -126,6 +126,16 @@ sut = GenericProvider('FakeProvider')
         'expected': datetime(2018, 1, 18, 17, 37, 33, tzinfo=tz.gettz('US/Eastern')),
         'timezone': 'US/Eastern',
         'fromtimestamp': True
+    },
+    {  # p22: hd-space test human date like 'yesterday at 12:00:00'
+        'pubdate': 'yesterday at {0}'.format((datetime.now() - timedelta(minutes=10, seconds=25)).strftime('%H:%M:%S')),
+        'expected': datetime.now().replace(microsecond=0, tzinfo=tz.gettz('UTC')) - timedelta(days=1, minutes=10, seconds=25),
+        'human_time': False
+    },
+    {  # p23: hd-space test human date like 'today at 12:00:00'
+        'pubdate': 'today at {0}'.format((datetime.now() - timedelta(minutes=10, seconds=25)).strftime('%H:%M:%S')),
+        'expected': datetime.now().replace(microsecond=0, tzinfo=tz.gettz('UTC')) - timedelta(days=0, minutes=10, seconds=25),
+        'human_time': False
     },
 ])
 def test_parse_pubdate(p):